# doc/api.yaml
openapi: 3.0.0
info: 
  description: WASAPhoto REST APIs
  title: WASAPhoto APIs
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
security:
  - BasicAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    userid:
      schema:
        description: unique id related to user
        type: integer
        example: 1
        readOnly: true
      name: userid
      in: path
      required: true
      description: User ID
    postid:
      schema:
        description: unique id related to the post
        type: integer
        example: 1
        readOnly: true
      name: postid
      in: path
      required: true
      description: Post ID
...
  responses:
    BadRequest:
      description: |-
        The request was not compliant with the documentation
        (eg. missing fields, etc).
    ForbiddenRequest:
      description: |-
        Clien executes not request to the contents without
        required permissions.    
    InternalServerError:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
    UnauthorizedRequest:
      description: |-
        The request had been executed without login
    NotFound:
      description: |-
        Component is not found
  schemas:
...
    PostChanged:
      title: Post
      description: "This object represent a single Post after some operation"
      type: object
      properties:
        postid:
          description: Unique post identifier.
          type: integer
          example: 1
          readOnly: true
        quantitylikes:
          description: likes quantity related to current post
          type: string
          example: '1k'
          pattern: '(^[0-9]{1,3}\.[0-9]{1}k$)|(^[0-9]{1,3}\.[0-9]{1}m$)|(^[0-9]{1,3}$)|(^[0-9]{1,3}k$)|(^[0-9]{1,3}m$)'
          minLength: 1
          maxLength: 6
        quantityDislikes:
          description: dislikes quantity related to current post
          type: string
          example: '1k'
          pattern: '(^[0-9]{1,3}\.[0-9]{1}k$)|(^[0-9]{1,3}\.[0-9]{1}m$)|(^[0-9]{1,3}$)|(^[0-9]{1,3}k$)|(^[0-9]{1,3}m$)'
          minLength: 1
          maxLength: 6
        
        
      example:
        {
          postid: 2,
          quantitylikes: '1k',
          quantityDislikes: '2.3k'
        }
    PostCreated:
      title: PostCreated
      description: "This object represents a single Post after operation create"
      type: object
      properties:
        postid:
          description: Unique post identifier.
          type: integer
          example: 1
          readOnly: true
      example:
          {
            postid: 2
          }
...
    PostToChange:
      title: PostToChange
      description: "This object represent a single Post after some operation"
      type: object
      properties:
        postid:
          description: Unique post identifier.
          type: integer
          example: 1
          readOnly: true
        text:
          description: Text related to post
          type: string
          minLength: 0
          maxLength: 255
          pattern: '^.*?$'
          example: I had been create the first post
        image:
          description: image related to post
          type: string
          format: binary
          minLength: 0
          maxLength: 100000
      example:
        {
          postid: 2,
          text: This is some post inside,
          image: "iVBORw0KGgoAAAANSUhEUgAAAgAAAAIACAIAAAB7GkOtAAAkmklEQVR4AezVgQAAAAACoP2lH+SSKAXgUgqAAAAQAAACAEAAAAgAAAEAIAAABACAAAAQAAACAEAAAAgAAAEAIAAABACAAAAQAAACAEAAAAgAAAEAIAAAAQAgAAAEAIAAABAAAAIAQAAACAAAAQAgAAAEAIAAABAAAAIAQAAACAAAAQAgAAAEAIAAABAAAAIAQAAAAgBAAAAIAAABACAAAAQAgAAA3gkAAAEAIAAABACAAAAQAAACAEAAAAgAAAEAIAAABACAAAAQAAACAEAAAAgAQAAACAAAAQAgAAAEAIAAABAAAAIAQAAACAAAAQAgAAAEAIAAABAAAAIAQAAACAAAAQAgAAAEAIAAABAAY+8c2CTX1jD6p67Gtu05tn3Gtm3btm3bZpJiquLs+11j3KyqrPW84wfTPT3rDfb+diETJhP+7RvuqWPO4f32nu321g3WupXZFQszC2dlZk02p4xJjx2aGtY32b9bst9vqUE9U8P7pUcPTo8fLn9kzpiYmTstu3SetWmNs3+Xe+aEf/Na8OypsrKq/AGgAADCMNA17+ol58Aea80yc/qE1JDeiR8/M7q21BpUfl7rj6URrV4Fo33j+LcfpQb2kJ6QUnGOHJCmCc20AqAAAErvot49ezK7emlqRP/4F11ftHz5Rm9dL/HT53IPYa1f6Z47HSYTCoACACgKvu/fuWXv2iqX9olu3xntGopk8yt6q7qJn780Z01yDu0L44YCoAAAXiN97+I5eUwv3nzxAj/fY3RsmhrQPbtysXfpvPJ9FXWAAgAIAnmOL29cE79/qzWuKqKMQuQjTXb/PrtqidzlqEgBFACAvDjNrliU7Pmj3rSGCDHK0ds0kNfX9raNYTymChWgAAC8G1fNmZNkoc6LHiRanT8nvv9UbgtkyakqDIACAPCuXZZ3uUaX5lj+LSPLnDKL5/j376p8BCgAAHm4b04fb3RqhtCLnNinneSeID+eDgEFACBPMDKzp8iiF/RdUtHqVkj2+sk5uCcnlw8BBQAQhs7xw/JeV55lo+xSit6yTnriSP/uLZULAAUAECbiso7T6Fx2j/hJ4pevZMyRlK4qF4ACAPAunE0N7qXVr4iRyyXyXl12loWmqQAoACgjXNfauDr2cXsUnCPbymQGUfDsiSpFgAIAcB1r7XKZiJlrEiQyqTQ9ckDw6IECoACghHFsmcSZ80PZWC/0F9lX7N+7owAoACgBbFvGNsjEgjzyIJuK5awbFgsBBQDFwLKyyxbIpPs89SB3A+lRgwJdUwAUALwDnpddvlBWnee7BInWqEpmztQwm1EAbywAADnmMPZeq0KSIJHbODnAUgW+AqAA4KX4D+7J8VuFKkES+6CNHKupAP67AADCjGlOHSdLCQtegkT27oUxXQFQAAJYW9ZH6k0v0ZvVkP0cTJKASBcAyPm08S/fi6YESfzzLt7VyypyAAUAlpUeP0yr/aeIS5ClouaMicp1FVAAEJ0hbkaXFv9QACGxj9p7168ooACgwHEdedn74sh+wpkzmbnTOHCGAoCCRR74ykLAVymAkNhnnf3bNxVQAFBQeJ4c0ygPfHEceX20BpWs9SsVUABQGPi3rsvZ4m+vAEKS/X7nnBkKAPKe7KolRdveRThuzLt2WQEFAHmJZaUG9kBkpMiRwz7lEAgFFADk3VQfWdtXIhYgPA5SlqWAAoC8wDm0V29aA3ORkkrsk47B08cqpwEKAILAnDmpxPf3EiKHQ7jnTqvcBCgACJOJxC9flZ4CCOfOs0KUAoAc3eRldGhc2gogJD1mSA4dLAMUADgH9mgNKuMmUjZJdvsuJ14LAwUActQfs31IGUemiIfxmCpHgAIAGfCAjEi5RAbKBo8eqHIAKAAI/NSI/miIlGP0VnXL/EgZoADAtpM9fkBApNxjtGukHFuVGUABsNwz/vUHqIfkwtnCZXcHABQAyIbM2PutUQ/B/hCtAgD//l29TQPUQ7A/RKsAQFZcGG2xP8H+ELECgODJIzb6EuwPkSsACJ4/Mzo1Qz0E+0O0CgACXTO6tkQ9BPtDtAoAwrjBmh+C/SGKBcB6f071ItgfIlcAEKZTsU87oR6C/SFiBQCuw15fgv0higUAqQHdUQ/B/hC5AoDMrMmoh2B/iFwBgL19E+oh2B8iVwDgnj+t1a+IfQj2h8gVAKN+9JZ1sA/B/hC9AmDRJxu+CPaHyBUA+H7ip8+xD8H+ELkCgPTIgdiHYH+IXAGAvWMz9iHYHyJXAOA/uKc1roaACPaHiBUAuE5pT/shsqxWhmknfv06Naxvetwwc+q4zNxp2WXzrXUr7K0b7L07JPITa82y7JK5mTlTzSlj0mMGp4b0Tvb+Of7le3qL2ti/5AEKAMRHCLoEo9X5swxPTQ3vl1kw0965xbtwNtCeqTBUxSM0Te/GVWf/ruzSedINiV+/Mto3xv5FBygAcPbvRtnFj7g42fdXuaJ3z55SVlaVFVIt9p4d6Qkj4p931ur+BfsDBQBvS/D0sd6sJvousrlkWJ5c4weGrnIBK+ueOp6ZPyPxy1davQrY/9UABQC+z6jnIkQe5csDevfMCeX7ubyhT94oJLp995ImwP5AAYA5bRw2f/sYnZtnFs7y799VeUWYSlpb1id+/1arWwH7KwEoAHBPHdNq/wmtvzFag0qpgT3c08dVnvO3Jti0JvZJR+wP0S4AsG2jUzPk/vrEPumQXb1UvFlw3X882e07qX/sD1EsADCnjsXvr4mssHTPnVYFjX/vdnrUIK1BZewPESoA8K5dftV6QZLs+aN35aKKDGE8lpk3XaZ/Y3+IQAFA4LPp96Vbt5L9fvdvXVeRJDRNWREge5WxfyFDAYCMGUD3L171+/fuKKU4CCjZ+xfsX5hQAOA/vP/fz3yJHH3jHD+s/guQ/Q2yUgj7FxoUAPz7sBeiN6meXb7w5Tu5IAxlwajeuh72LxAoALA2r8X7Eln+KGPawpiu4E17iZP9u2H/vIcCgDBuMPNHIrsf3POnFbw19raNcreE/fMYCgDSY4Zg/9Tw/mE2o94RCB4/in/7EfbPSygAkPE1EV/4r7eq6xzaq4oMBIEMGZWvIuyfZ1AAIEdKRXyVZxg3VLEB7/IFo0tz7J83UADgXjgT3fe99SpYf2XvLoDjOLqFDReEOfm/MDMzMzMzM+P9wswxhcHMbIvNIJmZmSUzyWJY3p45dyr9X5cSyzy7mul5nzrF6CQ+72a2d7pvlsDVb4ar3nuV7Q8C4A/BfOO/fuyTku97kUw6NxWz/UEAvC42elhAf+H16lPWhvWSMnButGf7gwB4mFLO3VVBPO3z/WcSiwlSLJY/0HnI5sXtDwKAaF5GAN/pFu7ZSZAuiVnTnN+XeGv7gwDADof0O36DM86Vh7GR+YL0UiuXJxez/UEAvCQx6qeSWy8J1MWN8YljBQABCDoVUqMPiXbfv+yBcwOx/e+6xrmzVwAQANgrf0sO29WZRP89Kl8+1ezt7zyDTs6fIwAIAMROqHHH6QDoqf366JKbjD3sr4oWCwACAIe9rpPe+3Un2vag0rsvMO+zv1q+VAAQAPzNVhPP2jQAzsRz96546ixznvvfcXVy7iwBQACg2SV5et3XP0N2q37/BCNOfF4WnzBGABAAbKSmXL6FAOgJ/3Zoya0X+ToA0cH9BAABwEZ2xSi94rc6sR77lT10jk+3f6RnZwFAAFCXNefRbQyAM4mBe1S+eorvtn/tXz8JAAKAf4gXqxF7bXsA9NR8d5SPTohWvfuqWEoAEADUZa9opnf69k603UGld53vg/vcH72jnou9ABAAqPGn7EgANp4QffpMj9/tlZw3WwAQAPyLXV6gV/mOz9Ddqj/07gnRSG5v2RQAAgBrziM7GQA94T8OLbntQs9d8NLoCwFAALCZr3/3dCUAzsR67Vf+sIdOiFY8/7BEo7IpAAQA9vKmene7NYmBu1e+dopHfvGbXDhP6gWAAEBNOM3dAOip/eHIkpsbOAChtn9J/QAQANRM1/s6FRPtcGDpPec35MOfRFzqBYAAwCr6NHUBcCaet1fFM2fw8AcwCgHg+P+2Tv5uNZ8cz8MfwHwEgOc/9U74z0NKbk/TCdGK5x7i4Q9AALAlVuEnaQuAM7GMfcsfOTsNAYhPmSAACAC2QI0/OZ0BcCYxaPfKN05O7RvfPn5bABAAbEn1VL2U0z+1jY8ouSlV3/1yze9WAAQAVuHHDRUAZyKdD0jFCdHa3xoLAAKALVMTTm/AADiT6Ltn5bNunhAtu/s6u7JCABAAbEl0hd7CDTwFu9Z8dhx3PQIgAOljr23nhQDoibQ4uHSnT4iWPXQbRz9dBhAArv9Nw8Qyd/aEaKRXZ9kaAAQAlhp9iIcCoGfQ7lVvnbSDH//vud4OhWQrABAAVE/RO9eDU9vk8JKbL9reAITaNZetAkAAYC9v5NkAOBPtckDZvedtx9n/26+0K8pkqwAQAKjp13s5APqEaMXzp3P231UAAYCqVSP28nQA9OQ7J0SP3frH/5svtdaulq0CQABgl+frDeuLibT6T+kdWzohWv35uwLAAASALwDqOyG6T/njZ20uAPExw2VbACAAsGbd56cA6Bm8e9U7J9Zz+vOBmyWZFADbggBAjT3KdwHQE2p2eMkt/zghWtv8ZwGAbQoAoiv1MvXpRLvuX3b/uRsDkCxcLACwLQGAvSHH1wFwJtF/z4qXTnO2f8UL94k3ASAA3AGZuneI1n55TKRnMwGAbQwA1PQbDQiAM2r4bhLfIKhj3GLl92GmLVMCApAathp1oCEBmHKZ/BNuaBTy+zC3NQvbAgKQCuFFensaMNayb8VEBIBZV2mL6wgA7NJ+xgRAqqeIiQgAM7FQiesIAOwVPxvy/GfMESK2mIgAMBmTEuI6AgBr4auGPP+Z/7wYigAwPw6ICwiAKW+Bdn/s4gwxFAFg3u4SFRAAXgKx2QOgyUoxFAFg7vklLC4jAFDVhnwBMPEsMRoBYCpCtriJAKBmmilfADwnRiMAzPw1lriIAMAu7mlGAOzVzcVoBIAZuUCJiwgAnF9OmfILgEliNALAZG56EhQEgDOgasSeYkXFaASAaZ7PSVACwEVgm4yafLGYjgAwX+fExEUEAGrK5QYEwFr4upiOADBvdIqKiwgA1LgTjfgGuIWYjgAwD/0RERcRAKiR+5oQgNIBYjoCwNzYOJy0xCUEAKrGkCNAobmCAASA2VBtizsIACKFhgRA1QgCEACmaIMl7iAAqBpvwhGg0YcIghEAZuYKS1xBAGCX5BpxBvQSQTACwIxZpMQVBAD22nYmnAGd87AgGAFgBs5KiisIAOzVzU0IwJL3BQEMAPeCgQBwGaS9opkgGAFg2o90KQAEAPayH0wIwKq/BMEIAPP74Li4ggDAWvqlCQFY20EQjAAw3+fFxBUEANaSD424CriXIBgBYD7PdCkABADW4neMeA9EP0EwAsB83IsAEIA6uAzALh8mCEYAmPd7RMUVBADW/OcMCIBUTRQEIwDMO11dCgABgDX3MRMCUDtLEIwAMG+6dyUAASAAjxMAHyEAzCvtowICUAePgCYIghEA5oW2EXEFAYC18DW+BPYRAsA804oAEIC6OAZa0lcQjAAwT7RwKQAEAFbhR/wQzEcIAPNUy4iAANTBqyDaC4IRAOb5Ni4FgADAXt6Il8H5CAFgXm7vUgAIAOyVv5gQgOVNBcEIAPN6x6iAAGyCC2HMRwCYt7tEBQSgDq6EfEgQjAAw73avJwAgAFwKbz4CwHzUy60AEABUjTchAKMPEQQjAMxnGTFxBwFApNCAADgjqlYQgAAwX2e7FQACAFVjSABCcwUBCADTrH9c3EIAoEbua8SlYAMEAQgA07LAvQAQAKjxJ5kQgNUtBAEIANN1bELcQgCgpl5hwknQha8LAhAAJm9qUtxCAGDNvt+Ik6AXCwIQAGbYPCUgAHVxL7wasadYMTEdAWAmFykBAXCLtew7Qw4CVU8SEADTZ/4aS+BWAGAX9zQjAPbq5gICYPqsKrMFrgUANdPNCIA1/3kBATB9KsMC9wIAVW1GANTEswUEwOi5pWnYFrgaAKixR5kQgOG7SbJSQAC4EBjbHgCo6dcb8jXAhkwBATB3/tstKnA3ALAWvmbK1wAvCAiAudO4b0xAALgYst5RY44QsQUEwNBpNzIhcDcAsEv7mREAZ6R6qoAAGDp9pycFLgcA4UXGBMBa9q2AABg6EwuVwO0AwFajDjQjAGrKZQICYOgs3WAJXA8A1PQbjfmfABVbLyAAJk4oJnA/ALAKPzFj+7frf2bPxf0FBMC4eeiPsCAVAYC9Icfvq7+iYK93cq+6qPeDjw95X+pAbVSYr7Jjfg/A+z2igpQEALFVvt7+84b8556s253tr2dRxTKpA3i8ecTvAfhjSFxAAP6FF0LkDjjxit7369Wv58fp7eX/AOUh24BHQH2mJQUpCgCsWff5bvWHh+3+XZ+L6q5+PTfmPZe0kvI3YOwiZUAAZq6wBCkKAOzljfy1/Vfl7/9k9o2bbn89w1dPlL8BrYfHDQhARcgWpCgAsMvzfbT9xww66vqMe/Sur3f+Z0wj+Rvwdpeo37f/vb+GBSkMAFStGrGX91d/Ytiurfqepbf8FuaSjIdX1fKDAEHSktuahf0eAKdhgoAHgPdCl+Xv/UbONXrFb3UaTW0tgYf5a6ytb1iOABEA2Msbe3n7zx588J1Zd+jlvi1zRdZjZdFKCTa0G5kwIABD5yQFqQ0Aqqd4dvtnDTjp8gx91nM75s/Z3STY8FTLiAEBWMld8KkPAGw1+hCvrf5Qwe5f9rlEL/Ttnetyng4lwhJUKCw24fnP3T+HBWkIAKy5j3lq+y8fesBjWTfv2PbX02lBrgQV2hvx/OeDnlFBGgIAe20772z/EYOOvjbz3h1e/Xpu7fNiTMUlkPBMKxOe/zgZE6QjAIiu8MLqjxfs+le/c5z17cp0XJAjwYOlGywz3gM6fokSpCcAUBNOb9jtX5K/z6s517qy+vVck/NUeaxKAgYdRyXMCEB5yBakJwCwCj9uwO0/Y/Aht2fe6eL21/PD1FYSJFCWPPqnCc9/Hm8eEaQvAKiZ1lDbv9eAUy7t/YDr21//MLiwaoUEBgrmKjM+/v80IC5IZwCgxp+c5tVfW7DHp3mX6WWdonlz1LcSGHi1Q9SMAIyYrwQGB4AbIouGHvhI1i16Tad0xq6bJgGAmSsM+fr3xsbhipAgvQFAzfS0bf/8gcdek6HPeqZ8Hhj4VtxKiOnwWUbMjAC81C4iSH8AoMafkurVHxu22699z9OrOW3z+6wuYjSsKrOdD85mBKDlsLgg/QGAVfRZSrd/cf4+L+dcl+btr78Nnl22SMyFXweZcP2LnslFSmBYAHgKNHXIobdm3pX+7a/nwYFvm/rbYKypsG9tasjH/1uahqMJQcMEAGrCaanY/l37n6bPejbg/Dazs5gIn2fGjPn4/99uUUFDBQD28qburv7qgj0/zL08RTudB0GYusyQs/96sicnBQ0WAMSL1Yg93dr+S4Ye9GDWrS6sb/ceBEWSUTEFLFuebxMxKQDFVbagAQMAa84jrmz/QQOPuyrjPu9sfz2fTvhVTIGcKUmTtr/zQzZBwwYAdnnBzp/1/LHvBXrhenB6LO4v/ofqiNz7a9ikAHQflxAQAF//IGBd/r7P59zgzdWv57LMR2aULBCfwy8DzTn6qWdFqS1o8ADAXtFsx7b/pMGH35z+s547dGNMabRCfAtjFyvDtv+zrSMCAuCZr4L32t7t37H/6Zdk6A3rg3lx+OfKVuJDKK2x7zPo4Y+edlwB5p0AwJrz6Lav/qr8Pd/NvVIvVh9N42ltxG9gi7zbXb/106hZUmwJPBIA2BWjtnH7Lxz6/+7Lus13219Pu/lZ4itwvik1b/u/yAvgvBYAqCmXb3X79xtwwpUZ9/t0++vJLcoXn8D8NdbNTcLmBSBzUlLgqQDALsnbwuqPFuzWqM9Fvl79G38hPHz1JPE81ETliRYR87a/kzQuAPBgAGCriWfVu/3XDN3vmZwbDdj+eq7Iemx6yTzxMCSUvNPVtEf/er7IjAk8GADY6zptuv3HDz7ixsy7jdn+eq7LeXpx5XLxJNgiX2XrN74ZOGMXKYEXAwA7ocYdV3f7t+l35sV6aRo3N+Y9N7+8SLwHvw+Jm7r97/8trCyBRwMAe+VvevVXFOz1du5VeleaOtfmPOW1Hwmjx3h97MfMaZ4fF3g3AFCh8IgD5g05+J6s283e/nquyn5iwvqZ4g0YOqdBX/fG6x8IAEbNaXJFb33WMxBzedZjXjgXhNELlT70aep81Csm8HgAEEpGbu7zQnAC4MylmY8MWD5SGg76z0jepO95N3cmL1UC7wcAmYWDAxUAZy7OeKjN3N4Cfu7L298CHgAoW90/8K1ABUDPB+OapfkSMTTP12d+DJ9+05MCvwQAw1dPDGAAnHl08LtrQsWC1FOWNOobC8L2d66yiSYEfgoAniv4NIAB0D8RmFw8R5BKtVH5uJf5219P2xEJgb8CgJmlC4IZAP21MHdJps7CtdbjzSMB2f63Ng2X1NgC3wUA749tFswA6Hlz1HclkXJxFbImJ29pqg/8BGJ+GxwX+DEAWFGz1vksHOQG3JD77NCV48QlvODzi8yYXovB+fi/odoW+DQAcC7SCnIA9Hw+8feaREiwcy/31499AjW/8/Hf1wFAeazq+txnaMAd/V6ZXDxbsP0iyegvMzpf3SLXfxucp/8EAH2XDScAej6b8GtxpEy2GUaumXxnv1f+/1s32v94feOyAH38HxIXGBAAvDria7b/xvfHdZifHbcSskVYFy55d0yTf9/I1vXt634qDMjH/1I+/psRAKyqXX9l1uNs/41z74A3nM+2Uh/EVLzTglynlPW/daPXk1f/Mdz4APw1NC4wJgDovDCPvb/JOdFvuVmsroSVzFgy6La+L239Ys7W7a5vVGPq9r/nl3B1RGBOAKBs64mh77P0N533xjaZV76E90flFuXrx/3bOJd2+uz6pmuNDEDW5KTAsABgQUWRuT8LcOFXY9NL5gfzU79zTOCe/q/vwD+0S3q8cM0v0w3b/s+0iihLYGAA8NvMzuz6LcxLw7+YuH5WcL7m/Wt295vynt+pF3H3fviqFlkmBWDCEiUwNQB8v7fVz3rMAwPf6rIwryxaKYYav36Gc8LnkoyH3fondln7ptc3MeGE6Ic9owKDAwDn91DO9Sls+a3OZf/L3jlwR7atXfhPtZn0sX1O27ZtM06+0TZj27Zt206qat33Wt9tJbVLzzPmcHfU6Tn3XuvFm99lwUBiS5beoFdWQU1/491izx+DdhhlAN/TnYtvlFm0+8tKy5oOgwICwLpxyXuCv39QF7FHwQvLrReq6quXdWm/hewx+s/qxfKFzpGWGwAuNlL6SQBQ8vER2wKQFMlczbqT3JJj/n1k8hXKnbbk1i/BuzT+Kc27fXvJ5X6Lc/8/XEeGbcf/CQD6PJkRNJV24gOJV/2qo5qH2s3K9LPai+Rhf0vM6XmeS015evboqP3VRssKgNRKnQJbCgDWRqZh5VOXVNHsSbgkZ+sprbl944Maj2kr7q6UIk6H3Iebok/N9VxqRmv6n69b7JBpKe5/0W9MAQFga1zLvouDT6/kilVGzsnZi7wfyGbKxsHWSf2kmjKDE8NyhSsfMKg21r3guTSvyTQLs7/MlwrR1+bv/j85DvcNK7C5AAA5NKA92NiSaku5PJAOg6PJt86lu13PvueS/1TeGJ6W+XtVhQfXxgXXxftUR76qCH5S6nen6I1b/rNbOQ9kkcOxFAf5Wz8GbZcTJwsuprp/yf5KpzkHQGShToHNBgBbwxZ5r8SmkREj8Ol2uxul5un+R1+NKbDlAAB5AsWkkHH1ctlCl3Bzc/8fHZj5TACAUhcz/w+TQsbWvDseSy71mU8AJFfoFBAAoDPoNFgag9DsR4ftrplFhajzPxZ+AQEAsjZdg6YhhKRCdJFDumndf+PdkQmdAgIA/onULH7qu9b4FoDQr/PdX5nK/b+6PlzXaVBAAMB/kNNRok0/EUJzHlywv9KhfQAE5kwqIADg/0Uq07WyAESF6LbF2laInvcdU0AAwFuQRla8SRuhmX+uEA3Txv033xsdm1RAAMA7kDZUzSwAobl33JZcNm6F6M9Ow+39BgUEALzPlAgtR0YjNPvxQbtr9UZy/8+vDhc06BUQAO8PhaErwg9qaAGICtE1ixzSjBEAAR968QsEAPSND/wRuk9TF0BUiHo8n173dwwdV0AAwEfQPdb3a8hubS0AUSF6zv5q+7S4/4Hnozq9AgIAPpLO0Z6fgnfiSloKzXq21e5G8RTdf8v9URY9EgAwVdpGur4P3KatBSAqRJcucAn5aPdf6THSM6SAAIBpQPbffhuwRXMXQFSIOttf7v1Q9//VeaSl16CAAIBpHBb0lf8mLEljoVlP9ttdr3t/9//+1nB1u14BAQDTvkHsS/+NmlsAokJ09SLHlPdx/y+vGa3kHwgAaBnuMMngaITmezxZcmnw7e6fVqVTRgUIAPoD6BNGJtGch2fsr7b9L/fPrDG++wMBAGO68T0Jl0xhAYgK0c2Lbxb+h/t/fX04p1avgAAAbdAZ9Ocz3E1gAQi9/GOBa9A/3P+bG8P59bg/AQA2Mzsaobl3He0v93x7c7iw0UTuDwQAeFaGzXr9G36EtNe3XmdKmnF/AgBMSkxT2nzPZfgR0lI/Bu2Q3hQFBACYnOLuKs1ahRFaE3m0d6xfAQEA5lMeuj32HN6EjK39iVekDk0BAQBmhd6gd8t/NvP1r5gUMpKuZN3RG/QKCAAwT+KaMuy8V2FVaHo1583vL8qDFBAAYObUD7awTQxNoz7zW5/dUaSAAACLYFQ3diLVEedCU9eqiMNtw50KCACwLOSdfc6bP7Aw9NE6k+46rp9QQACAJVLRW7cs7ABGhj5U0lwibYYKCACwaCb1kzI0Yvab3zE19J6SO6SqvnoFBABYS7NY5a8hu7E29E5dy75rbcc+QACA/K92zH30v2YHIfSp77r45kxlrQABALmdJTLIBbND/6G9CZc7R3sUEABg9UWi8pr/j55hxIN/WH2iAgIAbIeCrvLVEUewPxvXiVQnJrsRAGCj+NdEf+63Hh+0QX0TsDmxJUsBAQC2zNDkiFwO207LGJrrudQ578nw5KgCAgBAqBto3h1/EXO0hcveBna5AAEA/42cCfwUvBOXtEpJI0hKa64CIADgLZ3DT8v8pTIEx7SiOp+1L8oDdQadAiAA4H1KRZ+U+Vv6/TCy81l9t9jzg4/7gQAAGNONPy8P/NJ/I05qcVrotcK94PnAxJACIACmAjHwsiL4K/9NuKpFaIHXcofchz1U9wMBANM4SuhNZagUj+Ow5tzTe6fotVEau4AAAJjQT/pURSwNY+WkeUnmO8n4fnlXU8YGCACAvM7Sk6lO0lKE+ZpW66OORzemGZRBARAAoCVy2vCo1PeHoO0YsfblPdez71X21SkAAgBMiEEZklty9iVcYdmABloXdSywNta8TnuAAABoHe6U9ZPfBW7FpqddXwdslpFNbGoEAgAsYAOljBv7PnAbxj31wp4rWXdyOkoUAAEAlkVJT5VL/tMfuST4QH3iu/Zcultqa57OoFcABABYNKU91W75z9hG+XYtDdsvPyWpsNJbn+8DAQAghSsyXkKGEtt5r8LxRYu8V8r9uXdVeNtIlwIgAMAW0Bv0clXwuNR3V/wFGVxjU6Zv77NaTF/G7clPwKYPeYAAANAZdPmdZfeLvbbFnp3vucz6HF9W8MuihWMpDq8qgst7a///vi0AAgAIAyl2DK1LkNtjWVJmoUPopCVCFrCcSnOW867sjqKhiWH1oQAQAAB94wOZ7YUvyoPOZ7ivijhshq8Icp+xOuKI2L0M349oSC7rrTFKoxYAAQA0HneMdMvpeUxT2uuKEHlRkKlEW2JOS4mRUbNhnudSqWfdFH3qRKqjU95j2bEV2ZCc31XWNdqrAAgAAJPTNz4ohUbpbfnxzRnh9UkBNTEyJvNZWcC9Yk9ZkCKD8i9n3T6T7no0+daR5Bti5WfSXC5keEiz1Y2c+9JqK1WY8hQvJ/XBtXHxzZnSgSUfrW24c2hy5E/t1YEQAAAAAyF/6s3jryQaCAAAAQAgAAAEAIAAABAAAAIAQAAACAAAAQAgAAAEAIAAABAAAAIAQAAACAAAAQAgAAABACAAAAQAgAAAEAAAAgBAAAAIAAABACAAAAQAgAAAEAAAAgBAAAAIAAABACAAAAQAQCYAAAQAgAAAEAAAAgBAAAAIAAABACAAAAQAIAAABACAAAAQAAACAEAAAAgAAAEAIAAABACAAAAQAAACAEAAAAgAAAEAIAAABACAAAAQAAACAEAAAAgAQAAACAAAAQAgAAAEAIAAABAAAAIAQAAACAAAAQAgAAAEAIAAABAAAAIAQAAACAAAAQAgAAAEAIAAABAAAAdZJ8s6QoeY5gAAAABJRU5ErkJggg=="
        }
...
    Profile:
      title: Profile
      description: User represented as an object
      type: object
      properties:
        avatar:
          description: User's profile image
          type: string
          format: binary
        userid:
          description: Unique user identifier
          type: integer
          example: 1
          readOnly: true
        username:
          description: Unique username identifier
          type: string
          example: username1
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
        password:
          description: password to enter inside application
          type: string
          example: "*zafgz3lU341"
          pattern: '^.*?$'
          minLength: 8
          maxLength: 16
        quantitySubscribers:
          description: number of subscribers to the current account
          type: string
          example: "3k"
          pattern: '(^[0-9]{1,3}\.[0-9]{1}k$)|(^[0-9]{1,3}\.[0-9]{1}m$)|(^[0-9]{1,3}$)|(^[0-9]{1,3}k$)|(^[0-9]{1,3}m$)'
          minLength: 1
          maxLength: 6
        quantitySubscriptions:
          description: number of users current user subscribed to
          type: string
          example: "2k"
          pattern: '(^[0-9]{1,3}\.[0-9]{1}k$)|(^[0-9]{1,3}\.[0-9]{1}m$)|(^[0-9]{1,3}$)|(^[0-9]{1,3}k$)|(^[0-9]{1,3}m$)'
          minLength: 1
          maxLength: 6
          
      example:
        {
          avatar: "binary string",
          userid: 1,
          username: 'User1',
          quantitySubscribers: '2.3k',
          quantitySubscriptions: '34m'
        }
      
          
#there are some paths inside which u can check below          
paths:
...
  /profiles:
    get:
      operationId: searchForUser
      tags: 
        - Search
      summary: shows top users and permits to search for specific user
      description: |-
        Return information about current user from a database
      responses:
        "200":
          description: List of profiles in reverse order
          content:
            application/json:
              schema:
                description: returns list of profiles in reverse order
                type: array
                minItems: 0
                maxItems: 10
                items: { $ref: "#/components/schemas/Profile" }
                
              example: [
                  {
                      avatar: "binary string",
                      userid: 1,
                      username: 'User1',
                      quantitySubscribers: '2.9k',
                      quantitySubscriptions: '30m'
                  },
                  {
                      avatar: "binary string",
                      userid: 2,
                      username: 'another one',
                      quantitySubscribers: '2.3k',
                      quantitySubscriptions: '25m'
                  }
                  ]
              
              
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{userid}/posts:
    parameters:
      - $ref: "#/components/parameters/userid"
    get:
      operationId: getUserProfilePosts
      tags:
        - Posts
      summary: get home page of current user
      description: |-
        Return information about current user from a database
      responses:
        "200":
          description: List of posts in reverse order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      operationId: uploadPhoto
      tags:
        - Posts
      summary: create a post to share with others
      description: |-
        Create a post with a photo using the properties in the request body.
        The server will add photo to server
      security:
        - BearerAuth: []
      requestBody:
        description: Request for photo Upload
        required: true
        content:
        #below u can have data for uploading image maybe not application/json
          application/json:
            schema:
              $ref: '#/components/schemas/PostToChange'
      responses:
        "201":
          description: Post had been created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{userid}/posts/{postid}:
    parameters:
      - $ref: "#/components/parameters/userid"
      - $ref: "#/components/parameters/postid"
    delete:
      operationId: deletePhoto
      security:
        - BearerAuth: []
      tags:
        - Posts
      summary: delete a picture/post from profile
      description: |-
        Removes a post with a photo using the properties in the request body.
        The server will remove a photo
      
      responses:
        "204":
          description: Post had been deleted
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    put:
      operationId: modifyPhoto
      tags:
        - Posts
      summary: modify a picture/post under the profile
      description: |-
        Removes a post with a photo using the properties in the request body.
        The server will remove a photo
      requestBody:
        description: Requests for updating post
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostToChange'
      responses:
        "200":
          description: Post had been updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostChanged" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{userid}:
    parameters:
      - $ref: "#/components/parameters/userid"
    get:
      tags:
        - Account
      operationId: getUserProfile
      summary: shows all information related to certain account
      description: |-
        Show all posts of current user inside of application
      
      responses:
        "200":
          description: Posts of another user had been retrieved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedRequest" }
        "403": { $ref: "#/components/responses/ForbiddenRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...